1. Fixed Codacy issues.
